rename: redisassert -> serverassert
